Atl-fwd: Minor additions to ethtool dump implementation.
The patch incoroprates following changes,
 - Put ethtool-dump data structures in atl_dump.h for sharing it with parser.
 - Add driver/fw version to dump-data
 - Increment ATL_FWD_API_VERSION.
